Removing preview.
It never goes over well on imgur, even though that's what it was made for (to make the static preview of the first frame more informative). People seem to judge the post by the end frame (the preview frame) and downvote it whereas when the first frame is mostly black (e.g. the first frame) people almost always upvote the post, because they end up clicking into it and watching the time-lapse. That's my guess anyway.